% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distoverlap.R
\name{calc_drawsoverlap}
\alias{calc_drawsoverlap}
\title{Calculate Distribution overlap}
\usage{
calc_drawsoverlap(draws_mat)
}
\arguments{
\item{draws_mat}{A \code{draws_matrix}}
}
\value{
A n x n matrix
}
\description{
\code{calc_drawsoverlap} calculates the
generalized distribution overlap according to Weitzman (1970)
for a \code{draws_matrix}
}
