% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdr.R
\name{calc_hdr}
\alias{calc_hdr}
\title{Calculate the Highest Density Region}
\usage{
calc_hdr(vec_x, probs = c(0.5, 0.87, 0.99))
}
\arguments{
\item{vec_x}{A numeric vector containing the posterior samples.}

\item{probs}{A vector with the probabilities for the credible intervals.
The default calculates the 50\%, 87\%, and 99\% HDR.}
}
\value{
The output will be a matrix containing the credible intervals.
}
\description{
\code{calc_hdr} calculates the Highest Density Region (HDR)
from a vector containing posterior samples
}
