% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2data.R
\name{calc_meddelta_gammas}
\alias{calc_meddelta_gammas}
\title{Calculate median centered effects}
\usage{
calc_meddelta_gammas(draws_mat)
}
\arguments{
\item{draws_mat}{A draws matrix containing the effects}
}
\value{
The output will be a draws matrix with median centered effects
}
\description{
\code{calc_meddelta_gammas} centers the effects from a draws matrix
at the median
}
