% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distoverlap.R
\name{calc_overlap}
\alias{calc_overlap}
\title{Calculate Distribution overlap}
\usage{
calc_overlap(dist1, dist2)
}
\arguments{
\item{dist1}{An empirical distribution of equal length to dist2}

\item{dist2}{An empirical distribution of equal length to dist1}
}
\value{
A number
}
\description{
Utility function
}
