% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2data.R
\name{calc_prcteststat}
\alias{calc_prcteststat}
\title{Calculate test statistics for posterior retrodictive checks}
\usage{
calc_prcteststat(data)
}
\arguments{
\item{data}{A numerical vector or matrix}
}
\description{
Utility function
}
