% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stderrmeas.R
\name{calc_sem}
\alias{calc_sem}
\title{Calculate the standard error of measurement}
\usage{
calc_sem(sd, rel)
}
\arguments{
\item{sd}{A vector of standard deviations of measurement scales.}

\item{rel}{A vector of reliability coefficients of measurement scales.}
}
\value{
The output will be a numeric vector of equal length to input.
}
\description{
Utility function
}
