% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wassersteindistance.R
\name{calc_wasserstein}
\alias{calc_wasserstein}
\title{Calculate the Wasserstein distance}
\usage{
calc_wasserstein(x, y, type = c("l1", "l2", "energy"))
}
\arguments{
\item{x}{A distribution function as vector}

\item{y}{A distribution function as vector}

\item{type}{One of \code{"l1"}, \code{"l2"} or \code{"energy"}}
}
\value{
The output will be the distance
}
\description{
Utility function
}
