% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2graphs.R
\docType{data}
\name{colorscheme_frankmakrdiss}
\alias{colorscheme_frankmakrdiss}
\title{ggplot2 colorscheme}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 6 rows and 5 columns.
}
\usage{
colorscheme_frankmakrdiss
}
\description{
Utility data
}
\keyword{datasets}
