% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comms_data}
\alias{comms_data}
\title{Classification Scheme of Communities}
\format{
\subsection{\code{comms_data}}{

A data frame with 1,580 rows and 6 columns.
}

\subsection{Variables}{

\describe{
\item{person}{Person number from \code{comms_demo}}
\item{field}{Which of the 10 fields was used for the phrase}
\item{k_start}{The phrases before the classification process}
\item{k_iter1}{The phrases after iteration 1}
\item{k_iter2}{The phrases after iteration 2}
\item{k_iter3}{The phrases after iteration 3}
}
}
}
\source{
Primary data
}
\usage{
comms_data
}
\description{
Phrases for the communities in German
where persons have experienced themselves sense of community.
}
\keyword{datasets}
