% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conval_comms}
\alias{conval_comms}
\title{Community phrases for the Dataset conval_data}
\format{
\subsection{\code{commnames}}{

A data frame with 22 rows and 2 columns.
}

\subsection{Variables}{

\describe{
\item{long}{The original phrases used in the study material}
\item{short}{Short phrases used in text, tables, and graphs}
}
}
}
\source{
Primary data
}
\usage{
conval_comms
}
\description{
The phrases of the communities in German the persons could choose from
}
\keyword{datasets}
