% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conval_data}
\alias{conval_data}
\title{Content validity of tests for psychological sense of community}
\format{
\subsection{\code{contentvalidity}}{

A data frame with 5,609 rows and 79 columns.
}

\subsection{Variables}{

\describe{
\item{person}{Person number from \code{conval_demo}}
\item{comm}{The specific community from \code{conval_comms}
for which the test items were rated}
\item{value}{The grade to which the community is a
community of shared values}
\item{purpose}{The grade to which the community is a
community of purpose}
\item{value_x_purpose}{Interaction variable}
\item{chavis_[1-25]}{The grade to which the test item [1-25] from
Chavis et al. (2008) is necessary to characterize sense of community}
\item{omoto_[1-18]}{The grade to which the test item [1-18] from
Omoto and Snyder (2010) is necessary to characterize sense of community}
\item{jason_[1-9]}{The grade to which the test item [1-9] from
Jason et al. (2015) is necessary to characterize sense of community}
\item{halamova_[1-22]}{The grade to which the test item [1-22] from
Halamová et al. (2018) is necessary to characterize sense of community}
}
}
}
\source{
Primary data
}
\usage{
conval_data
}
\description{
The content validity of the items of 4 different tests
was rated in specific community contexts.
}
\keyword{datasets}
