% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2data.R
\name{count_bins_iter}
\alias{count_bins_iter}
\title{Count bins from a draws matrix}
\usage{
count_bins_iter(draws_mat, bx)
}
\arguments{
\item{draws_mat}{The draws matrix \code{y_rep_mat}}

\item{bx}{The breaking points}
}
\description{
Utility function
}
