% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distmat}
\alias{distmat}
\alias{chavis_dist}
\alias{omoto_dist}
\alias{jason_dist}
\alias{halamova_dist}
\title{Distance Matrices of Community Effects}
\format{
\subsection{\code{chavis_dist}}{

A named list of length 6 containing 22 x 22 distance matrices.
}

\subsection{\code{omoto_dist}}{

A named list of length 10 containing 22 x 22 distance matrices.
}

\subsection{\code{jason_dist}}{

A named list of length 4 containing 22 x 22 distance matrices.
}

\subsection{\code{halamova_dist}}{

A named list of length 4 containing 22 x 22 distance matrices.
}
}
\source{
Primary data.
}
\usage{
chavis_dist

omoto_dist

jason_dist

halamova_dist
}
\description{
The Matrices contain the pairwise energy distances
of the distributions of community effects in the ratings.
The item ratings are aggregated at the test and the dimension level.
}
\references{
Chavis, D. M., Lee, K. S., & Acosta, J. D. (2008).
\emph{The Sense of Community Index (SCI) revised:}
\emph{The reliability and validity of the SCI-2.}
Paper presented at the 2nd International Community Psychology Conference.
Lisboa, Portugal.
Retrieved October 17, 2018, from \url{https://senseofcommunity.com}.

Halamová, J. Kanovsky, M., & Naništová, E. (2018).
Development and psychometric analysis of the
sense of community descriptors scale.
\emph{Psychological Intervention, 27(1), 44-55.}
\url{https://doi.org/10.5093/pi2018a8}.

Jason, L. A., Stevens, E., & Ram, D. (2015).
Development of a three-factor psychological sense of community scale.
\emph{Journal of Community Psychology, 43(8), 973-985.}
\url{https://doi.org/10.1002/jcop.21726}.

Omoto, A. M., & Snyder, M. (2010).
Influences of psychological sense of community
on voluntary helping and prosocial action.
In S. Stürmer & M. Snyder (Eds.),
\emph{The psychology of prosocial behavior:}
\emph{Group processes, intergroup relations, and helping (pp. 223-243).}
Wiley-Blackwell.
\url{https://doi.org/10.1002/9781444307948.ch12}.
}
\keyword{datasets}
