% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanmodels.R
\name{fit_test}
\alias{fit_test}
\title{Example function for cmdstanr integration}
\usage{
fit_test(stan_data)
}
\arguments{
\item{stan_data}{The data used for the model}
}
\value{
The function returns a stan fit object.
}
\description{
Test function
}
