% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frankmakrdiss-package.R
\docType{package}
\name{frankmakrdiss-package}
\alias{frankmakrdiss}
\alias{frankmakrdiss-package}
\title{frankmakrdiss: Accompanying Package for Frank Krumbholz's dissertation}
\description{
Makes reproduction, replication, or adaptation of the dissertations analyses easier.
}
\author{
\strong{Maintainer}: Frank Krumbholz \email{frankmakr@mailbox.org} [copyright holder]

}
\keyword{internal}
