% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2data.R
\name{make_compjitter}
\alias{make_compjitter}
\title{Make plot data for showing the pairwise overlapping probabilities}
\usage{
make_compjitter(probs_list, dims_list)
}
\arguments{
\item{probs_list}{A list containing the
pairwise overlapping probabilities}

\item{dims_list}{A list containing the names of the test dimensions}
}
\value{
The ouput will be a data frame in the format for
using the corresponding \code{plot_compjitter} function
}
\description{
\code{make_compjitter} prepares the data
in the format for the function \code{plot_compjitter}.
}
