% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2data.R
\name{make_dp_prcjitter}
\alias{make_dp_prcjitter}
\title{Plot Data for a Posterior Retrodictive Check}
\usage{
make_dp_prcjitter(y, y_rep_mat)
}
\arguments{
\item{y}{A vector containing the sample replications}

\item{y_rep_mat}{A draws matrix containing the population replications}
}
\value{
The ouput will be a data frame in the format for
using the corresponding \code{plot_prcjitter} function
}
\description{
\code{make_dp_prcjitter} prepares the data
in the format for the function \code{plot_dp_prcjitter}.
}
