% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2data.R
\name{make_dp_prcstats}
\alias{make_dp_prcstats}
\title{Plot Data for a Posterior Retrodictive Check}
\usage{
make_dp_prcstats(y, y_rep_mat, tikzdevice = FALSE)
}
\arguments{
\item{y}{A vector containing the sample replications}

\item{y_rep_mat}{A draws matrix containing the population replications}

\item{tikzdevice}{Logical indicator for preparing the text for tikzDevice
which defaults to \code{tikzdevice = FALSE}}
}
\value{
The ouput will be a data frame in the format for
using the corresponding \code{plot_dp_prcstats} function
}
\description{
\code{make_dp_prcstats} prepares the data
in the format for the function \code{plot_dp_prcstats}.
}
