% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2data.R
\name{make_gammapointrange}
\alias{make_gammapointrange}
\title{Plot Data for Community Effects}
\usage{
make_gammapointrange(draws_mat_1, draws_mat_2, probs = 0.87)
}
\arguments{
\item{draws_mat_1}{A \code{draws_matrix} containing the community effects
from the unadjusted model}

\item{draws_mat_2}{A \code{draws_matrix} containing the community effects
from the adjusted model}

\item{probs}{The probability for the highest density region
which defaults to \code{probs = 0.87}}
}
\value{
A data frame in the format for
using the corresponding \code{plot_gammapointrange()} function
}
\description{
\code{make_gammapointrange()} prepares the data
in the format for the function \code{plot_gammapointrange()}.
}
\details{
A \code{draws_matrix} is a numeric matrix
in which the rows are posterior draws
and the columns are variables.
}
