% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2data.R
\name{make_hypo_prcstats}
\alias{make_hypo_prcstats}
\title{Plot Data for a Posterior Retrodictive Check}
\usage{
make_hypo_prcstats(y, y_rep_mat)
}
\arguments{
\item{y}{A numeric vector containing the observations}

\item{y_rep_mat}{A \code{draws_matrix} containing the replicated observations}
}
\value{
A data frame in the format for
using the corresponding \code{plot_hypo_prcstats()} function
}
\description{
\code{make_hypo_prcstats()} prepares the data
in the format for the function \code{plot_hypo_prcstats()}.
}
\details{
A \code{draws_matrix} is a numeric matrix
in which the rows are posterior draws
and the columns are variables.
}
