% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2data.R
\name{make_lambdapointrange}
\alias{make_lambdapointrange}
\title{Plot Data for Cluster Probabilities}
\usage{
make_lambdapointrange(draws_mat_list, labels, probs = 0.87, tikzdevice = FALSE)
}
\arguments{
\item{draws_mat_list}{A list of draws matrices containing the lambdas}

\item{labels}{A character vector containing the names of the graph facets}

\item{probs}{A numeric value representing
the probability for the highest density region,
which defaults to \code{probs = 0.87}}

\item{tikzdevice}{Logical indicator for preparing the text for tikzDevice
which defaults to \code{tikzdevice = FALSE}}
}
\value{
The ouput will be a data frame in the format for
using the corresponding \code{plot_dp_prcstats} function
}
\description{
\code{make_lambdapointrange} prepares the data
in the format for the function \code{plot_lambdapointrange}.
}
