% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2data.R
\name{make_logclusterdensities}
\alias{make_logclusterdensities}
\title{Plot Data for Cluster Densities}
\usage{
make_logclusterdensities(draws_mat_list)
}
\arguments{
\item{draws_mat_list}{A named list of draws matrices containing the
log cluster densities}
}
\value{
The ouput will be a data frame in the format for
using the corresponding \code{plot_logclusterdensities} function
}
\description{
\code{make_logclusterdensities} prepares the data
in the format for the function \code{plot_logclusterdensities}.
}
