% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2data.R
\name{make_mds_plotdata}
\alias{make_mds_plotdata}
\title{Plot Data for MDS Configuration}
\usage{
make_mds_plotdata(draws_mat_list, target_mat)
}
\arguments{
\item{draws_mat_list}{A named list of \code{draws_matrix} objects
containing the MDS configurations}

\item{target_mat}{A \code{draws_matrix}
containing the target configuration for the procrustes transformation}
}
\value{
A data frame in the format for
using the corresponding \code{plot_dp_mds()} function
}
\description{
\code{make_mds_plotdata()} prepares the data
in the format for the function \code{plot_dp_mds()}.
}
\details{
A \code{draws_matrix} is a numeric matrix
in which the rows are posterior draws
and the columns are variables.
}
