% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2data.R
\name{make_prcjitter}
\alias{make_prcjitter}
\title{Plot Data for a Posterior Retrodictive Check}
\usage{
make_prcjitter(y, y_rep_mat)
}
\arguments{
\item{y}{A list containing the sample replications}

\item{y_rep_mat}{A list containing the population replications}
}
\value{
The ouput will be a data frame in the format for
using the corresponding \code{plot_prcjitter} function
}
\description{
\code{make_prcjitter} prepares the data
in the format for the function \code{plot_prcjitter}.
}
