% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2data.R
\name{make_prcstats}
\alias{make_prcstats}
\title{Make plot data for a posterior retrodictive check}
\usage{
make_prcstats(y, y_rep_mat)
}
\arguments{
\item{y}{A list containing the sample replications}

\item{y_rep_mat}{A list containing the population replications}
}
\value{
The ouput will be a data frame in the format for
using the corresponding \code{plot_prcstats} function
}
\description{
\code{make_prcstats} prepares the data
in the format for the function \code{plot_prcstats}.
}
