% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2data.R
\name{make_samplepointrange}
\alias{make_samplepointrange}
\title{Plot Data for Evaluating Sample Representativity}
\usage{
make_samplepointrange(sample_simlist, population_simlist, probs = 0.87)
}
\arguments{
\item{sample_simlist}{A list containing the sample replications}

\item{population_simlist}{A list containing the population replications}

\item{probs}{The probability of the highest density region,
which defaults to \code{probs = 0.87}}
}
\value{
The ouput will be a data frame in the format for
using the corresponding \code{plot_samplepointrange} function
}
\description{
\code{make_samplepointrange} prepares the data
in the format for the function \code{plot_samplepointrange}.
}
