% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2data.R
\name{make_sigmapointrange}
\alias{make_sigmapointrange}
\title{Make plot data for showing a variance component}
\usage{
make_sigmapointrange(draws_mat_1, draws_mat_2, probs = 0.87)
}
\arguments{
\item{draws_mat_1}{A draws matrix containing the
superpopulation standard deviation from the unadjusted model}

\item{draws_mat_2}{A draws matrix containing the
superpopulation standard deviation from the adjusted model}

\item{probs}{A numeric value representing
the probability for the highest density region,
which defaults to \code{probs = 0.87}}
}
\value{
The ouput will be a data frame in the format for
using the corresponding \code{plot_sigmapointrange} function
}
\description{
\code{make_sigmapointrange} prepares the data
in the format for the function \code{plot_sigmapointrange}.
}
