% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2data.R
\name{make_thetapointrange}
\alias{make_thetapointrange}
\title{Plot data for community probabilities}
\usage{
make_thetapointrange(
  draws_mat,
  probs = 0.87,
  n_group_tilde,
  n_sample_tilde,
  tikzdevice = FALSE
)
}
\arguments{
\item{draws_mat}{A \code{draws_matrix} containing the
community probabilities}

\item{probs}{The probability for the highest density region
which defaults to \code{probs = 0.87}}

\item{n_group_tilde}{Expected group size in the future sample}

\item{n_sample_tilde}{Expected size of the future sample}

\item{tikzdevice}{Logical indicator for preparing the text for \code{tikzDevice}
which defaults to \code{tikzdevice = FALSE}}
}
\value{
A data frame in the format for
using the corresponding \code{plot_thetapointrange()} function
}
\description{
\code{make_thetapointrange()} prepares the data
in the format for the function \code{plot_thetapointrange()}.
}
\details{
A \code{draws_matrix} is a numeric matrix
in which the rows are posterior draws
and the columns are variables.
}
