% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2graphs.R
\name{plot_compjitter}
\alias{plot_compjitter}
\title{Plot Pairwise Distribution Overlapping}
\usage{
plot_compjitter(compjitter)
}
\arguments{
\item{compjitter}{A data frame returned from the function
\code{make_compjitter}}
}
\value{
The output will be a ggplot2 graph
}
\description{
\code{plot_compjitter} plots data
in the format from the function \code{make_compjitter}.
}
