% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2graphs.R
\name{plot_dp_prcjitter}
\alias{plot_dp_prcjitter}
\title{Plot Posterior Retrodictive Check}
\usage{
plot_dp_prcjitter(prcjitter)
}
\arguments{
\item{prcjitter}{A data frame returned from the function
\code{make_dp_prcjitter}}
}
\value{
The output will be a ggplot2 graph
}
\description{
\code{plot_dp_prcjitter} plots data
in the format from the function \code{make_dp_prcjitter}.
}
