% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2graphs.R
\name{plot_dp_prcstats}
\alias{plot_dp_prcstats}
\title{Plot Posterior Retrodictive Check}
\usage{
plot_dp_prcstats(prcstats)
}
\arguments{
\item{prcstats}{A data frame returned from the function
\code{make_dp_prcstats}}
}
\value{
A \code{ggplot2} graph
}
\description{
\code{plot_dp_prcstats} plots data
in the format from the function \code{make_dp_prcstats}.
}
