% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2graphs.R
\name{plot_gammapointrange}
\alias{plot_gammapointrange}
\title{Plot Community Effects}
\usage{
plot_gammapointrange(gammapointrange, tikzdevice = FALSE)
}
\arguments{
\item{gammapointrange}{A data frame returned from the function
\code{make_gammapointrange}}

\item{tikzdevice}{Logical indicator for preparing the text for tikzDevice
which defaults to \code{tikzdevice = FALSE}}
}
\value{
The output will be a ggplot2 graph
}
\description{
\code{plot_gammapointrange} plots data
in the format from the function \code{make_gammapointrange}.
}
