% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2graphs.R
\name{plot_hypopointrange}
\alias{plot_hypopointrange}
\title{Plot Hypotheses Testing}
\usage{
plot_hypopointrange(hypopointrange, type, tikzdevice = FALSE)
}
\arguments{
\item{hypopointrange}{A data frame returned from the function
\code{make_hypopointrange()}}

\item{type}{One of \code{"color"} or \code{"single"}}

\item{tikzdevice}{Logical indicator for preparing the text for \code{tikzDevice}
which defaults to \code{tikzdevice = FALSE}}
}
\value{
A \code{ggplot2} graph
}
\description{
\code{plot_hypopointrange()} plots data
in the format from the function \code{make_hypopointrange()}.
}
