% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2graphs.R
\name{plot_lambdapointrange}
\alias{plot_lambdapointrange}
\title{Plot Cluster Probabilities}
\usage{
plot_lambdapointrange(lambdapointrange, tikzdevice = FALSE)
}
\arguments{
\item{lambdapointrange}{A data frame returned from the function
\code{make_lambdapointrange}}

\item{tikzdevice}{Logical indicator for preparing the text for \code{tikzDevice}
which defaults to \code{tikzdevice = FALSE}}
}
\value{
A \code{ggplot2} graph
}
\description{
\code{plot_lambdapointrange} plots data
in the format from the function \code{make_lambdapointrange}.
}
