% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2graphs.R
\name{plot_logclusterdensities}
\alias{plot_logclusterdensities}
\title{Plot Cluster Densities}
\usage{
plot_logclusterdensities(lcddata, tikzdevice = FALSE)
}
\arguments{
\item{lcddata}{A data frame returned from the function
\code{make_logclusterdensities()}}

\item{tikzdevice}{Logical indicator for preparing the text for \code{tikzDevice}
which defaults to \code{tikzdevice = FALSE}}
}
\value{
A \code{ggplot2} graph
}
\description{
\code{plot_logclusterdensities()} plots data
in the format from the function \code{make_logclusterdensities()}.
}
