% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2graphs.R
\name{plot_mds}
\alias{plot_mds}
\title{Plot MDS Configuration}
\usage{
plot_mds(mds_plotdata, mds_color_groups)
}
\arguments{
\item{mds_plotdata}{A data frame returned from the function
\code{make_mds_plotdata}}

\item{mds_color_groups}{A numerical vector of \code{length= 4}
containing the numbers of the communities in \code{conval_comms}
to be highlighted}
}
\value{
The output will be a ggplot2 graph
}
\description{
\code{plot_mds} plots data
in the format from the function \code{make_mds_plotdata}.
}
