% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2graphs.R
\name{plot_prcjitter}
\alias{plot_prcjitter}
\title{Plot Posterior Retrodictive Check}
\usage{
plot_prcjitter(prcjitter)
}
\arguments{
\item{prcjitter}{A data frame returned from the function
\code{make_prcjitter}}
}
\value{
A \code{ggplot2} graph
}
\description{
\code{plot_prcjitter} plots data
in the format from the function \code{make_prcjitter}.
}
