% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2graphs.R
\name{plot_samplepointrange}
\alias{plot_samplepointrange}
\title{Plot Sample Characteristics}
\usage{
plot_samplepointrange(samplepointrange, tikzdevice = FALSE)
}
\arguments{
\item{samplepointrange}{A data frame returned from the function
\code{make_samplepointrange}}

\item{tikzdevice}{Logical indicator for preparing the text for \code{tikzDevice}
which defaults to \code{tikzdevice = FALSE}}
}
\value{
A \code{ggplot2} graph
}
\description{
\code{plot_samplepointrange} plots data
in the format from the function \code{make_samplepointrange}.
}
