% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2graphs.R
\name{plot_thetapointrange}
\alias{plot_thetapointrange}
\title{Plot Community Probabilities}
\usage{
plot_thetapointrange(thetapointrange)
}
\arguments{
\item{thetapointrange}{A data frame returned from the function
\code{make_thetapointrange}}
}
\value{
The output will be a ggplot2 graph
}
\description{
\code{plot_thetapointrange} plots data
in the format from the function \code{make_thetapointrange}.
}
