% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2data.R
\name{qgroups_iter}
\alias{qgroups_iter}
\title{Group quantiles from a counts and a quantile matrix}
\usage{
qgroups_iter(counts_mat, q_mat)
}
\arguments{
\item{counts_mat}{A matrix returned from \code{count_bins_iter}}

\item{q_mat}{The row quantiles from a matrix returned from
\code{count_bins_iter}}
}
\description{
Utility function
}
