% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samples}
\alias{samples}
\alias{comms_demo}
\alias{conval_demo}
\title{Samples of the studies}
\format{
\subsection{\code{comms_demo}}{

A data frame with 158 rows and 7 columns.
}

\subsection{\code{conval_demo}}{

A data frame with 1,465 rows and 7 columns.
}

\subsection{Variables}{

\describe{
\item{person}{Person number}
\item{gender}{Gender coded as male = 1, female = 2, divers = 3}
\item{age}{Age in years}
\item{state}{One of the 16 German states the person is living
in alphabetical order
or 17 = living outside from Germany}
\item{studstatus}{Student status used in the FernUniversität
official statistics}
\item{noexpert}{Knowledge of a theory of psychological sense of community
coded as 0 = Knowledge, 1 = No knowledge}
\item{numcomm}{The number of communities the person mentioned}
}
}
}
\source{
Primary data
}
\usage{
comms_demo

conval_demo
}
\description{
The sample characteristics
}
\keyword{datasets}
