% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablebody.R
\name{sep_after}
\alias{sep_after}
\title{Add LaTeX tabular row seperation}
\usage{
sep_after(x, y = character())
}
\arguments{
\item{x}{The argument \code{space.after} from \code{make_table_body}}

\item{y}{A necessary dummy variable}
}
\description{
Utility function
}
