% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2data.R
\name{sim_sampleprobs}
\alias{sim_sampleprobs}
\title{One Replication of Sample Probabilities}
\usage{
sim_sampleprobs(n, probs)
}
\arguments{
\item{n}{The sample size}

\item{probs}{A numeric vector containing the probabilities}
}
\value{
A numeric matrix containing the sample probabilities
}
\description{
Utility function
}
