% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanmodels.R
\name{stan_dissmodel}
\alias{stan_dissmodel}
\title{Fit Doctoral Dissertation's Stan Models}
\usage{
stan_dissmodel(model_number, stan_data, ...)
}
\arguments{
\item{model_number}{A character string containing the number
from the doctoral dissertation (i.e. "12.2")}

\item{stan_data}{The data used for the model as a named list}

\item{...}{Other arguments for \code{sample()}}
}
\value{
The function returns a stan fit object.
}
\description{
Fits the stan models from the doctoral dissertation
}
\details{
The compiled model is stored in the R user cache.
}
