% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaaic.R
\name{tabulate_paic}
\alias{tabulate_paic}
\title{Tabulate the probability in pairwise relative AIC comparison}
\usage{
tabulate_paic(data)
}
\arguments{
\item{data}{A numeric vector with AIC values}
}
\value{
A matrix with pairwise comparisons
containing the probability minimizing the information loss
}
\description{
\code{tabulate_paic} calculates the probability
that one model minimizes the information loss over another model.
}
