% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustes.R
\name{transform_procrustes}
\alias{transform_procrustes}
\title{Procrustes Transformation of MDS Configuration}
\usage{
transform_procrustes(x_mat, y_mat, dilation = FALSE)
}
\arguments{
\item{x_mat}{A numerical matirix representing the MDS target configuration}

\item{y_mat}{A numerical matrix representing the MDS configuration
to be compared}

\item{dilation}{A logical indicator for implementing optional scaling
which defaults to \code{dilation = FALSE}}
}
\value{
A numerical matrix representing the transformed MDS configuration
}
\description{
\code{transform_procrustes} calculates the transformation of a
multidimensional scaling configuration Y
to a target configuration X.
}
