% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wassersteindistance.R
\name{calc_drawswasserstein}
\alias{calc_drawswasserstein}
\title{Wasserstein Distance}
\usage{
calc_drawswasserstein(draws_mat, ...)
}
\arguments{
\item{draws_mat}{A \code{draws_matrix} containing the distributions}

\item{...}{The type of distance calculated,
which must be one of \code{"l1"}, \code{"l2"} or \code{"energy"}}
}
\value{
The output will be a distance matrix
}
\description{
\code{calc_drawswasserstein} calculates the
L1 or L2 Wasserstein distance or the Energy distance
for all pairwise combinations of distributions in a \code{draws_matrix}
}
