% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdr.R
\name{calc_hdr}
\alias{calc_hdr}
\title{Calculate the Highest Density Region}
\usage{
calc_hdr(vec_x, probs = c(0.5, 0.87, 0.99))
}
\arguments{
\item{vec_x}{A numeric vector containing the posterior samples}

\item{probs}{A numeric vector with the probabilities for the credible intervals
which defaults to \code{probs = c(0.50, 0.87, 0.99)}}
}
\value{
A named numeric matrix containing the credible intervals
}
\description{
\code{calc_hdr()} calculates the highest density region (HDR)
from a numeric vector containing posterior samples
}
