% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2data.R
\name{calc_meddelta_gammas}
\alias{calc_meddelta_gammas}
\title{Median Centered Effects}
\usage{
calc_meddelta_gammas(draws_mat)
}
\arguments{
\item{draws_mat}{A \code{draws_matrix} containing the effects}
}
\value{
A \code{draws_matrix} with median centered effects
}
\description{
\code{calc_meddelta_gammas()} centers the effects from a \code{draws_matrix}
at the median
}
\details{
A \code{draws_matrix} is a numeric matrix
in which the rows are posterior draws
and the columns are variables.
}
