% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolingfactors.R
\name{calc_poolstats}
\alias{calc_poolstats}
\title{Calculation of the pooling factors for a multilevel model}
\usage{
calc_poolstats(y, draws, pars = c("gamma", "delta", "y"))
}
\arguments{
\item{y}{A vector with the response variable}

\item{draws}{A draws matrix with the relevant pooling parameters.}

\item{pars}{A character vector with the names of the relevant residuals
of the multilevel model.}
}
\value{
The output will be a vector containing the pooling factors.
}
\description{
\code{calc_poolfac} calculates the pooling factors
according to Gelman and Pardoe (2006)
from a cmdstanr fit object
}
