% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stderrmeas.R
\name{calc_semci}
\alias{calc_semci}
\title{Calculate the standard error of measurement and confidence intervals}
\usage{
calc_semci(data, sd, rel, se0 = 2.8, p_ci = NA, dist = "none", ...)
}
\arguments{
\item{data}{A vector containing the measured values}

\item{sd}{A vector of standard deviations of measurement scales.}

\item{rel}{A vector of reliability coefficients of measurement scales.}

\item{se0}{The number of standard errors from 0 which are regarded
as the critical value for the confidence interval}

\item{p_ci}{The confidence niveau in percent}

\item{dist}{One of \code{"normal"}, \code{"t"}, or \code{"none"}}

\item{...}{The number of degrees of freedom for the student t distribution}
}
\value{
The output will be a data.frame containing the inputs,
the standard error of measurement, and the lower and upper bounds
of the confidence interval
}
\description{
\code{calc_semci} returns the standard error of measurement
from a given reliability and the corresponding standard deviation.
The confidence intervals are calculated using
a normal or a student t distribution
or according to the Tschebyschow inequality for an unknown distribution.
}
