% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conval_data}
\alias{conval_data}
\title{Content Validity of Tests for Psychological Sense of Community}
\format{
\subsection{\code{contentvalidity}}{

A data frame with 5,609 rows and 79 columns.
}

\subsection{Variables}{

\describe{
\item{person}{Person number from \code{conval_demo}}
\item{comm}{The specific community from \code{conval_comms}
for which the test items were rated}
\item{value}{The grade to which the community is a
community of shared values}
\item{purpose}{The grade to which the community is a
community of purpose}
\item{value_x_purpose}{Interaction variable of \code{value} and \code{purpose}}
\item{chavis_[1-25]}{The grade to which the test item [1-25] from
Chavis et al. (2008)
is rated necessary to characterize sense of community}
\item{omoto_[1-18]}{The grade to which the test item [1-18] from
Omoto and Snyder (2010)
is rated necessary to characterize sense of community}
\item{jason_[1-9]}{The grade to which the test item [1-9] from
Jason et al. (2015)
is rated necessary to characterize sense of community}
\item{halamova_[1-22]}{The grade to which the test item [1-22] from
Halamová et al. (2018)
is rated necessary to characterize sense of community}
}
}
}
\source{
Primary data
}
\usage{
conval_data
}
\description{
The rated content validity of the items of 4 different tests
in specific community contexts.
}
\references{
Chavis, D. M., Lee, K. S., & Acosta, J. D. (2008).
\emph{The Sense of Community Index (SCI) revised:}
\emph{The reliability and validity of the SCI-2.}
Paper presented at the 2nd International Community Psychology Conference.
Lisboa, Portugal.
Retrieved October 17, 2018, from \url{https://senseofcommunity.com}.

Halamová, J. Kanovsky, M., & Naništová, E. (2018).
Development and psychometric analysis of the
sense of community descriptors scale.
\emph{Psychological Intervention, 27(1), 44-55.}
\url{https://doi.org/10.5093/pi2018a8}.

Jason, L. A., Stevens, E., & Ram, D. (2015).
Development of a three-factor psychological sense of community scale.
\emph{Journal of Community Psychology, 43(8), 973-985.}
\url{https://doi.org/10.1002/jcop.21726}.

Omoto, A. M., & Snyder, M. (2010).
Influences of psychological sense of community
on voluntary helping and prosocial action.
In S. Stürmer & M. Snyder (Eds.),
\emph{The psychology of prosocial behavior:}
\emph{Group processes, intergroup relations, and helping (pp. 223-243).}
Wiley-Blackwell.
\url{https://doi.org/10.1002/9781444307948.ch12}.
}
\keyword{datasets}
