% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2data.R
\name{make_gammapointrange}
\alias{make_gammapointrange}
\title{Plot Data for Community Effects}
\usage{
make_gammapointrange(draws_mat_1, draws_mat_2, probs = 0.87)
}
\arguments{
\item{draws_mat_1}{A draws matrix containing the community effects
from the unadjusted model}

\item{draws_mat_2}{A draws matrix containing the community effects
from the adjusted model}

\item{probs}{A numeric value representing
the probability for the highest density region,
which defaults to \code{probs = 0.87}}
}
\value{
The ouput will be a data frame in the format for
using the corresponding \code{plot_gammapointrange} function
}
\description{
\code{make_gammapointrange} prepares the data
in the format for the function \code{plot_gammapointrange}.
}
