% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2data.R
\name{make_prcjitter}
\alias{make_prcjitter}
\title{Plot Data for a Posterior Retrodictive Check}
\usage{
make_prcjitter(y, y_rep_mat)
}
\arguments{
\item{y}{A nnumeric vector containing the sample replications}

\item{y_rep_mat}{A \code{draws_matrix} containing the population replications}
}
\value{
A data frame in the format for
using the corresponding \code{plot_prcjitter()} function
}
\description{
\code{make_prcjitter()} prepares the data
in the format for the function \code{plot_prcjitter()}.
}
\details{
A \code{draws_matrix} is a numeric matrix
in which the rows are posterior draws
and the columns are variables.
}
