% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2data.R
\name{make_prcstats}
\alias{make_prcstats}
\title{Plot Data for a Posterior Retrodictive Check}
\usage{
make_prcstats(y, y_rep_mat, tikzdevice = FALSE)
}
\arguments{
\item{y}{A numeric vector containing the observations}

\item{y_rep_mat}{A \code{draws_matrix} containing the replicated observations}

\item{tikzdevice}{Logical indicator for preparing the text for \code{tikzDevice}
which defaults to \code{tikzdevice = FALSE}}
}
\value{
A data frame in the format for
using the corresponding \code{plot_prcstats()} function
}
\description{
\code{make_prcstats()} prepares the data
in the format for the function \code{plot_prcstats()}.
}
\details{
A \code{draws_matrix} is a numeric matrix
in which the rows are posterior draws
and the columns are variables.
}
