% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2data.R
\name{make_sigmapointrange}
\alias{make_sigmapointrange}
\title{Plot Data for Variance components}
\usage{
make_sigmapointrange(
  draws_mat_1,
  draws_mat_2,
  probs = 0.87,
  tikzdevice = FALSE
)
}
\arguments{
\item{draws_mat_1}{A \code{draws_matrix} containing the
superpopulation standard deviation from the unadjusted model}

\item{draws_mat_2}{A \code{draws_matrix} containing the
superpopulation standard deviation from the adjusted model}

\item{probs}{The probability for the highest density region
which defaults to \code{probs = 0.87}}

\item{tikzdevice}{Logical indicator for preparing the text for \code{tikzDevice}
which defaults to \code{tikzdevice = FALSE}}
}
\value{
A data frame in the format for
using the corresponding \code{plot_sigmapointrange()} function
}
\description{
\code{make_sigmapointrange()} prepares the data
in the format for the function \code{plot_sigmapointrange()}.
}
\details{
A \code{draws_matrix} is a numeric matrix
in which the rows are posterior draws
and the columns are variables.
}
