% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablebody.R
\name{make_table_body}
\alias{make_table_body}
\title{LaTeX Tabular Body}
\usage{
make_table_body(
  data,
  space.after = 5,
  digits = 4,
  cor.tab = FALSE,
  row.names = TRUE
)
}
\arguments{
\item{data}{A named matrix or data.frame containing the table data}

\item{space.after}{A numeric vector or single number
indicating the number of rows after which row additional space is inserted}

\item{digits}{The number of digits to be displayed in the table}

\item{cor.tab}{Is the table a correlation table}

\item{row.names}{Does the data contain row.names}
}
\value{
The LaTeX table body
}
\description{
\code{make_table_body} strips the table body
from a \code{knitr::kable} table
}
