% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2data.R
\name{make_thetapointrange}
\alias{make_thetapointrange}
\title{Plot data for community probabilities}
\usage{
make_thetapointrange(
  draws_mat,
  probs = 0.87,
  n_group_tilde,
  n_sample_tilde,
  tikzdevice = FALSE
)
}
\arguments{
\item{draws_mat}{A draws matrix containing the
community probabilities}

\item{probs}{The probability for the Highest Density Region
which defaults to \code{probs = 0.87}}

\item{n_group_tilde}{Expected group size in the future sample}

\item{n_sample_tilde}{Expected size of the future sample}

\item{tikzdevice}{Logical indicator for preparing the text for tikzDevice
which defaults to \code{tikzdevice = FALSE}}
}
\value{
The output will be a data frame in the format for
using the corresponding \code{plot_thetapointrange} function
}
\description{
\code{make_thetapointrange} prepares the data
in the format for the function \code{plot_thetapointrange}.
}
