% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2graphs.R
\name{plot_compjitter}
\alias{plot_compjitter}
\title{Plot Pairwise Distribution Overlapping}
\usage{
plot_compjitter(compjitter, tikzdevice = FALSE)
}
\arguments{
\item{compjitter}{A data frame returned from the function
\code{make_compjitter()}}

\item{tikzdevice}{Logical indicator for preparing the text for \code{tikzDevice}
which defaults to \code{tikzdevice = FALSE}}
}
\value{
A \code{ggplot2} graph
}
\description{
\code{plot_compjitter()} plots data
in the format from the function \code{make_compjitter()}.
}
