% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2graphs.R
\name{plot_samplepointrange}
\alias{plot_samplepointrange}
\title{Plot sample characteristics}
\usage{
plot_samplepointrange(samplepointrange)
}
\arguments{
\item{samplepointrange}{A data frame returned from the function
\code{make_samplepointrange}}
}
\value{
The output will be a ggplot graph
}
\description{
\code{plot_samplepointrange} plots data
in the format from the function \code{make_samplepointrange}.
}
