% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2graphs.R
\name{plot_sigmapointrange}
\alias{plot_sigmapointrange}
\title{Plot Variance Components}
\usage{
plot_sigmapointrange(sigmapointrange, tikzdevice = FALSE)
}
\arguments{
\item{sigmapointrange}{A data frame returned from the function
\code{make_sigmapointrange()}}

\item{tikzdevice}{Logical indicator for preparing the text for \code{tikzDevice}
which defaults to \code{tikzdevice = FALSE}}
}
\value{
A \code{ggplot2} graph
}
\description{
\code{plot_sigmapointrange()} plots data
in the format from the function \code{make_sigmapointrange()}.
}
