% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2graphs.R
\name{plot_thetapointrange}
\alias{plot_thetapointrange}
\title{Plot Community Probabilities}
\usage{
plot_thetapointrange(thetapointrange, tikzdevice = FALSE)
}
\arguments{
\item{thetapointrange}{A data frame returned from the function
\code{make_thetapointrange()}}

\item{tikzdevice}{Logical indicator for preparing the text for \code{tikzDevice}
which defaults to \code{tikzdevice = FALSE}}
}
\value{
A \code{ggplot2} graph
}
\description{
\code{plot_thetapointrange()} plots data
in the format from the function \code{make_thetapointrange()}.
}
