% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samples}
\alias{samples}
\alias{comms_demo}
\alias{conval_demo}
\title{Sample Characteristics}
\format{
\subsection{\code{comms_demo}}{

A data frame with 158 rows and 7 columns.
}

\subsection{\code{conval_demo}}{

A data frame with 1,465 rows and 7 columns.
}

\subsection{Variables}{

\describe{
\item{person}{Person number}
\item{gender}{Gender coded as 1 = Male, 2 = Female, 3 = Divers}
\item{age}{Age in years}
\item{state}{One of the 16 German states the person is living
in alphabetical order
or 17 = living outside from Germany}
\item{studstatus}{Student status used in the FernUniversität
official statistics (Berichtswesen, 2022, 2023)}
\item{noexpert}{Formal theoretical knowledge of a theory of
psychological sense of community coded as 0 = Knowledge, 1 = No knowledge}
\item{numcomm}{The number of communities the person mentioned}
}
}
}
\source{
Primary data

Berichtswesen. (2022, March 25).
\emph{Studierendenstatistik für Sommersemester 2018}
[Student statistics for summer semester 2018].
FernUniversität in Hagen.
\url{https://www.fernuni-hagen.de/uniintern/organisation/statistik/semesterstatistik/sose2018.shtml}.

Berichtswesen. (2023, March 31).
\emph{Studierendenstatistik für Wintersemester 2019/20}
[Student statistics for winter semester 2019/20].
FernUniversität in Hagen.
\url{https://www.fernuni-hagen.de/uniintern/organisation/statistik/semesterstatistik/wise2019-20.shtml}.
}
\usage{
comms_demo

conval_demo
}
\description{
The characteristics of the samples for the studies.
}
\keyword{datasets}
