% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanmodels.R
\name{stan_dissmodel}
\alias{stan_dissmodel}
\title{Fit Doctoral Dissertation's Stan Models}
\usage{
stan_dissmodel(model_number, stan_data, ...)
}
\arguments{
\item{model_number}{A character string containing the model number
from the doctoral dissertation, which must be one of
\code{"12.2"}, \code{"12.3"}, \code{"13.1"}, \code{"13.3"}, \code{"13.6"}, \code{"13.6_adj"}, \code{"13.8"}, or \code{"14"}.}

\item{stan_data}{The data used for the model as a named list
with the names corresponding to the variables in model description
from the doctoral dissertation}

\item{...}{Other arguments for \code{cmdstanr::cmdstan_model()} or
\code{cmdstanr::cmdstan_sample()}}
}
\value{
A \code{CmdStanMCMC} object
}
\description{
\code{stan_dissmodel()} fits the Stan models from the doctoral dissertation.
}
\details{
The compiled model is stored in the R user cache
\code{tools::R_user_dir(package = "frankmakrdiss", which = "cache")}.
}
